Fix relative imports in moved tests
