+++ renderers/angular/a2ui_explorer/src/app/tests/v0_9/22_credit-card.spec.ts
@@ -15,7 +15,7 @@
  */
 
 import {ComponentFixture} from '@angular/core/testing';
-import {DemoComponent} from '../demo.component';
+import {DemoComponent} from '../../demo.component';
 import {getCanvas, loadExample} from '../test_utils';
 
 describe('Example: Credit Card', () => {
